structure

+++ src/app/app.component.ts
@@ -1,30 +1,51 @@
 import { HttpClientModule } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router, RouterOutlet } from '@angular/router';
 import { LoadingComponentComponent } from './loading-component/loading-component.component';
 import { AuthService } from './services/auth.service';
+import { interval, Subscription, switchMap } from 'rxjs';
+import { SessionTimeoutComponent } from './session-timeout/session-timeout.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, HttpClientModule, LoadingComponentComponent],
+  imports: [
+    RouterOutlet,
+    HttpClientModule,
+    LoadingComponentComponent,
+    SessionTimeoutComponent,
+  ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })
-export class AppComponent implements OnInit {
+export class AppComponent implements OnInit, OnDestroy {
   title = 'Shayyikli-System';
+  showTimeout = false;
+  private tokenCheckSub!: Subscription;
 
   constructor(private authService: AuthService, private router: Router) {}
+
   ngOnInit(): void {
-    this.authService.validateToken().subscribe({
-      next: (res) => {
-        console.log('Token validation successful:', res);
-      },
-      error: (err) => {
-        localStorage.removeItem('authToken');
-        localStorage.removeItem('userRoles');
-        sessionStorage.clear();
-      },
-    });
+    if (
+      localStorage.getItem('authToken') &&
+      localStorage.getItem('userRoles')
+    ) {
+      this.tokenCheckSub = interval(1)
+        .pipe(switchMap(() => this.authService.validateToken()))
+        .subscribe({
+          next: () => {},
+          error: () => {
+            this.tokenCheckSub.unsubscribe();
+            localStorage.removeItem('authToken');
+            localStorage.removeItem('userRoles');
+            sessionStorage.clear();
+            this.showTimeout = true;
+          },
+        });
+    }
+  }
+
+  ngOnDestroy(): void {
+    this.tokenCheckSub?.unsubscribe();
   }
 }
